Added missing ids in AgregateRoots
