+++ src/core/AccountPhone/domain/AccountPhone.ts
@@ -1,13 +1,18 @@
 import { AccountId } from '../../Account/domain/value-object/AccountId';
 import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
-import { Phone } from '../../Shared/domain/Phone/Phone';
+// import { Phone } from '../../Shared/domain/Phone/Phone';
 import { Primitives } from '../../Shared/domain/Primitives';
 import { Disabled } from '../../Shared/domain/value-object/Disabled';
+import { PhoneNumber } from '../../Shared/domain/value-object/PhoneNumber';
+import { PhonePrefix } from '../../Shared/domain/value-object/PhonePrefix';
+import { PhoneId } from './value-object/PhoneId';
 
 export class AccountPhone extends AggregateRoot {
   constructor(
-    readonly phone: Phone,
     readonly accountId: AccountId,
+    readonly id: PhoneId,
+    readonly number: PhoneNumber,
+    readonly prefix: PhonePrefix,
     readonly disabled: Disabled
   ) {
     super();
@@ -15,17 +20,21 @@ export class AccountPhone extends AggregateRoot {
 
   static fromPrimitives(plainData: Primitives<AccountPhone>): AccountPhone {
     return new AccountPhone(
-      Phone.fromPrimitives({
-        id: plainData.phone.id,
-        number: plainData.phone.number,
-        prefix: plainData.phone.prefix,
-      }),
       new AccountId(plainData.accountId),
+      new PhoneId(plainData.id),
+      new PhoneNumber(plainData.number),
+      new PhonePrefix(plainData.prefix),
       new Disabled(plainData.disabled)
     );
   }
 
   toPrimitives(): Primitives<AccountPhone> {
-    throw new Error('Method not implemented.');
+    return {
+      accountId: this.accountId.value,
+      id: this.id.value,
+      number: this.number.value,
+      prefix: this.prefix.value,
+      disabled: this.disabled.value,
+    };
   }
 }
