+++ src/core/AccountPhone/domain/value-object/PhoneId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class PhoneId extends Uuid {}
