+++ src/core/Contact/domain/Contact.ts
@@ -1,7 +1,14 @@
 import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
+import { PhoneNumber } from '../../Shared/domain/value-object/PhoneNumber';
+import { PhonePrefix } from '../../Shared/domain/value-object/PhonePrefix';
+import { ContactId } from './value-object/ContactId';
 
 export class Contact extends AggregateRoot {
-  constructor() {
+  constructor(
+    readonly id: ContactId,
+    readonly number: PhoneNumber,
+    readonly prefix: PhonePrefix
+  ) {
     super();
   }
