+++ src/core/Message/domain/Message.ts
@@ -1,7 +1,8 @@
 import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
+import { MessageId } from './value-object/MessageId';
 
 export class Message extends AggregateRoot {
-  constructor() {
+  constructor(readonly id: MessageId) {
     super();
   }
