--- src/core/Shared/domain/Phone/value-object/PhoneId.ts
@@ -1,3 +1,0 @@
-import { Uuid } from '../../value-object/Uuid';
-
-export class PhoneId extends Uuid {}
