+++ src/core/Shared/domain/value-object/PhoneNumber.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from './StringValueObject';
+
+export class PhoneNumber extends StringValueObject {}
